+++ src/lights/include/ILight.hpp
@@ -22,7 +22,7 @@ class ILight {
 public:
     virtual ~ILight() = default;
 
-    [[nodiscard]] virtual LightSample sample(const Point3f &p) const = 0;
+    [[nodiscard]] virtual LightSample sample(const Point3d &p) const = 0;
 
     [[nodiscard]] virtual bool isDelta() const = 0;
 };
